feat(core): give themes their own URLs (/themes and /themes/:id)
Themes browsing was hidden behind ?f=__themes__&theme=<id> on the home
route. Lifts the sidebar + mobile chrome into a shared HomeShell layout
route, exposes folder data via useOutletContext, and adds /themes plus
/themes/:themeId as standalone routes. Selecting Themes in the sidebar
now navigates instead of mutating query params, and links to a specific
theme can be shared as plain URLs.
